+++ tests/fields/test_datetime.py
@@ -1,8 +1,13 @@
 import os
+import warnings
 from datetime import datetime
 
+import pytest
+
 from tests.common import DummyPostData
+from wtforms.fields import DateField
 from wtforms.fields import DateTimeField
+from wtforms.fields import DateTimeLocalField
 from wtforms.form import Form
 
 
@@ -18,24 +23,30 @@ class F(Form):
     )
 
 
+def make_datetime_form(*args, **kwargs):
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        return F(*args, **kwargs)
+
+
 def test_basic():
     d = datetime(2008, 5, 5, 4, 30, 0, 0)
     # Basic test with both inputs
-    form = F(DummyPostData(a=["2008-05-05", "04:30:00"]))
+    form = make_datetime_form(DummyPostData(a=["2008-05-05", "04:30:00"]))
     assert form.a.data == d
     assert (
         form.a()
         == """<input id="a" name="a" type="datetime" value="2008-05-05 04:30:00">"""
     )
 
-    form = F(DummyPostData(b=["2008-05-05 04:30"]))
+    form = make_datetime_form(DummyPostData(b=["2008-05-05 04:30"]))
     assert form.b.data == d
     assert (
         form.b()
         == """<input id="b" name="b" type="datetime" value="2008-05-05 04:30">"""
     )
 
-    form = F(DummyPostData(c=["5/5/2008 4:30"]))
+    form = make_datetime_form(DummyPostData(c=["5/5/2008 4:30"]))
     assert form.c.data == d
     assert (
         form.c() == """<input id="c" name="c" type="datetime" value="5/5/2008 4:30">"""
@@ -43,11 +54,11 @@ def test_basic():
     assert form.validate()
 
     # Test with a missing input
-    form = F(DummyPostData(a=["2008-05-05"]))
+    form = make_datetime_form(DummyPostData(a=["2008-05-05"]))
     assert not form.validate()
     assert form.a.errors[0] == "Not a valid datetime value."
 
-    form = F(a=d, b=d, c=d)
+    form = make_datetime_form(a=d, b=d, c=d)
     assert form.validate()
     assert form.a._value() == "2008-05-05 04:30:00"
     assert form.b._value() == "2008-05-05 04:30"
@@ -57,7 +68,9 @@ def test_basic():
 def test_microseconds():
     d = datetime(2011, 5, 7, 3, 23, 14, 424200)
     F = make_form(a=DateTimeField(format="%Y-%m-%d %H:%M:%S.%f"))
-    form = F(DummyPostData(a=["2011-05-07 03:23:14.4242"]))
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        form = F(DummyPostData(a=["2011-05-07 03:23:14.4242"]))
     assert d == form.a.data
 
 
@@ -65,8 +78,30 @@ def test_multiple_formats():
     d = datetime(2020, 3, 4, 5, 6)
     F = make_form(a=DateTimeField(format=["%Y-%m-%d %H:%M", "%Y%m%d%H%M"]))
 
-    form = F(DummyPostData(a=["2020-03-04 05:06"]))
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        form = F(DummyPostData(a=["2020-03-04 05:06"]))
     assert d == form.a.data
 
-    form = F(DummyPostData(a=["202003040506"]))
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        form = F(DummyPostData(a=["202003040506"]))
     assert d == form.a.data
+
+
+def test_datetimefield_warns():
+    F = make_form(a=DateTimeField())
+
+    with pytest.warns(
+        DeprecationWarning,
+        match=r"DateTimeField.*DateTimeLocalField.*WTForms 3\.4",
+    ):
+        F()
+
+
+def test_datetimefield_subclasses_do_not_warn():
+    F = make_form(a=DateField(), b=DateTimeLocalField())
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("error", DeprecationWarning)
+        F()